turno-computadora.js: se crea y tiene como función realizar el turno d ela computadora
