+++ src/blackjack/index.js
@@ -10,7 +10,7 @@ import _ from 'underscore';
 // import { pedirCarta } from "./usecases/pedir-carta";
 // import { valorCarta } from "./usecases/valor-carta";
 /** Simplificamos y organizamos las exportaciones en el archivo index.js */
-import { crearDeck, pedirCarta, acumularPuntos, crearCarta, determinarGanador } from "./usecases/index";
+import { crearDeck, pedirCarta, acumularPuntos, crearCarta, turnoComputadora } from "./usecases/index";
 
 /**
  * Son funciones anonimas auto-invocadas
@@ -62,29 +62,6 @@ const miModulo = (() => {
       btnDetener.disabled = false;
   }
 
-
-
-  // Turno de la Computadora
-  const turnoComputadora = ( ) => {
-
-
-      let [ puntosMinimos, puntosComputadora ] = puntosJugadores; // Desestructuración de arreglos.
-
-      do{
-          const carta = pedirCarta( deck );
-          deck = deck.filter((i) => i !== carta); // Filtramos y eliminamos la carta seleccionada.
-          const turnoComputadora = puntosJugadores.length - 1;
-
-          puntosComputadora = acumularPuntos( carta, turnoComputadora, puntosJugadores );
-          crearCarta( carta, turnoComputadora );
-
-      } while( (puntosComputadora < puntosMinimos) && (puntosMinimos <= 21));
-
-      determinarGanador( puntosMinimos, puntosComputadora );
-  }
-
-
-
   // Eventos 
   btnPedir.addEventListener( 'click',() => {
 
@@ -99,7 +76,7 @@ const miModulo = (() => {
           console.warn('Lo siento mucho, perdiste');
           btnPedir.disabled = true; // deshabilita el boton
           btnDetener.disabled = true;
-          turnoComputadora( puntosJugador );
+          turnoComputadora( puntosJugadores, deck );
 
       } else if ( puntosJugador === 21 ){
           console.warn('21, genial');
@@ -116,7 +93,7 @@ const miModulo = (() => {
       btnPedir.disabled = true;
       btnDetener.disabled = true;
       
-      turnoComputadora( puntosJugadores[0] );
+      turnoComputadora( puntosJugadores, deck );
   });
 
   // Evento boton Nuevo
