+++ src/blackjack/usecases/index.js
@@ -4,3 +4,4 @@ export { valorCarta } from "./valor-carta";
 export { acumularPuntos }  from "./acumular-puntos";
 export { crearCarta } from "./crear-carta";
 export { determinarGanador } from "./determinar-ganador";
+export { turnoComputadora } from "./turno-computadora";
